fix(fam65xx): correct 65C816 native mode stack pointer handling
Fix 65C816 stack pointer to use full 16-bit addressing in native mode
instead of being forced to page 1. In native mode (E=0), SP can address
anywhere in bank 0 (0x0000-0xFFFF), while emulation mode and other 6502
family processors keep SP on page 1 (0x0100-0x01FF).

Changes include:
- Remove ZBR usage for stack operations; always use bank 0
- Add runtime check in address resolution for SP in native vs emulation
- Update register initialization to set full 16-bit SP value
- Fix BRK instruction to decrement full 16-bit SP in native mode
- Update test harness to handle 16-bit SP values for 65C816
- Change SP type from uint8_t to uint16_t in relevant interfaces

+++ code/cpp/tests/json_parser.h
@@ -42,7 +42,7 @@ typedef struct {
 // CPU state structure matching ProcessorTests format
 typedef struct {
     uint16_t pc;
-    uint8_t s;     // stack pointer (always 8-bit)
+    uint16_t s;    // stack pointer (8-bit for 6502, 16-bit for 65816 native mode)
     uint16_t a;    // accumulator (8-bit for 6502, 16-bit for 65816)
     uint16_t x;    // X register (8-bit for 6502, 16-bit for 65816)
     uint16_t y;    // Y register (8-bit for 6502, 16-bit for 65816)
